Add Testing capabilities to ConfigServer.py script

+++ config_server/ConfigServer.py
@@ -50,9 +50,11 @@ import urlparse
 import tempfile
 import shutil
 import json
+
 try:
     from requests.packages.urllib3.exceptions import InsecureRequestWarning
     requests.packages.urllib3.disable_warnings(InsecureRequestWarning)
+    urllib3.disable_warnings()
 except:
     pass
 def load(input_file, section=None):
@@ -64,8 +66,9 @@ def load(input_file, section=None):
     for prop, val in parser.items(section):
         dict[prop]=val
     return dict
-def enable_feed_action(cb_server_url, api_token, config, ssl_verify=True ):
-
+def enable_feed_action(cb_server_url, api_token, config, ssl_verify=True, test=None ):
+    if test:
+        print "----Validating configuration against Feeds----"
     feeds = requests.get(urlparse.urljoin(cb_server_url, '/api/v1/feed'),
                          headers={'X-Auth-Token': api_token}, verify=ssl_verify)
     feeds.raise_for_status()
@@ -74,8 +77,11 @@ def enable_feed_action(cb_server_url, api_token, config, ssl_verify=True ):
         fid=f['id']
         if f['enabled'] == False:
             f['enabled']=True
-            data = json.dumps(f)
-            requests.put(urlparse.urljoin(cb_server_url, '/api/v1/feed/{0}'.format(fid)),
+            if test:
+                print "%s Feed not enabled" % (fn)
+            else:
+                data = json.dumps(f)
+                requests.put(urlparse.urljoin(cb_server_url, '/api/v1/feed/{0}'.format(fid)),
                           data=data, headers={'X-Auth-Token': api_token}, verify=ssl_verify)
         if config.has_key(fn):
             f_actions = requests.get(urlparse.urljoin(cb_server_url, '/api/v1/feed/{0}/action'.format(fid)),
@@ -89,13 +95,36 @@ def enable_feed_action(cb_server_url, api_token, config, ssl_verify=True ):
                             if int(c) == int(f['action_type']):
                                 found=True
                     if found == False:
-                        print "Applying Feed Configuration for %s with %s" % (fn, c)
-                        data = json.dumps({'action_type': int(c), 'watchlist_id': None, 'action_data':'{"email_recipients":[1]}', 'group_id':int(fid)})
-                        requests.post(urlparse.urljoin(cb_server_url, '/api/v1/feed/{0}/action'.format(fid)),
+                        if test:
+                            print "%s Feed not configed for a type of %s" % (fn, c)
+                        else:
+                            print "Applying Feed Configuration for %s with %s" % (fn, c)
+                            data = json.dumps({'action_type': int(c), 'watchlist_id': None, 'action_data':'{"email_recipients":[1]}', 'group_id':int(fid)})
+                            requests.post(urlparse.urljoin(cb_server_url, '/api/v1/feed/{0}/action'.format(fid)),
                                       data=data, headers={'X-Auth-Token': api_token}, verify=ssl_verify)
+def report_watchlist_action(cb_server_url, api_token,  ssl_verify=True):
+    
+    print "----Reporting all watchlists configured to alert----"
+    watchlists = requests.get(urlparse.urljoin(cb_server_url, '/api/v1/watchlist'),
+                         headers={'X-Auth-Token': api_token}, verify=ssl_verify)
+    watchlists.raise_for_status()
+    for w in watchlists.json():
+        wn=w['name']
+        wid=w['id']
+        w_actions = requests.get(urlparse.urljoin(cb_server_url, '/api/v1/watchlist/{0}/action'.format(wid)),
+                                     headers={'X-Auth-Token': api_token}, verify=ssl_verify)
+        if w_actions.status_code == 200:
+            found=False 
+            for wa in w_actions.json():
+                if 3 == int(wa['action_type']):
+                    found=True
+            if found == True:
+                print "%s Watchlist configed for Alert" % (wn)
 
-def configure_groups(cb_server_url, api_token, config , ssl_verify=True):
+def configure_groups(cb_server_url, api_token, config , ssl_verify=True, test=None):
 
+    if test:
+        print "----Validating configuration against Groups----"
     groups = requests.get(urlparse.urljoin(cb_server_url, '/api/group'),
                          headers={'X-Auth-Token': api_token}, verify=ssl_verify)
     groups.raise_for_status()
@@ -104,17 +133,23 @@ def configure_groups(cb_server_url, api_token, config , ssl_verify=True):
         if int(config['banning']) == 1 and g['banning_enabled'] == False:
             g['banning_enabled']=True
             update=True
+            if test:
+                print "%s Group not configured for banning" % (g['name'])
         if int(config['tamper']) == 1 and int(g['tamper_level']) == 0:
             g['tamper_level']=1
             update=True
-        if update == True:
+            if test:
+                print "%s Group not configured for tamper" % (g['name'])
+        if update == True and not test:
             print "Updating Tamper and/or Banniing on Group %s" % (g['name'])
             data = json.dumps(g)
             requests.put(urlparse.urljoin(cb_server_url, '/api/group/{0}'.format(g['id'])),
                           data=data, headers={'X-Auth-Token': api_token}, verify=ssl_verify)
 
-def enable_data_sharing(cb_server_url, api_token, config ,ssl_verify=True):
+def enable_data_sharing(cb_server_url, api_token, config ,ssl_verify=True, test=None):
 
+    if test:
+        print "----Validating configuration against Sharing Settings----"
     groups = requests.get(urlparse.urljoin(cb_server_url, '/api/group'),
                          headers={'X-Auth-Token': api_token}, verify=ssl_verify)
     groups.raise_for_status()
@@ -127,12 +162,18 @@ def enable_data_sharing(cb_server_url, api_token, config ,ssl_verify=True):
                 if "TICEVT" == a['what']:
                     found=True
             if found == False:
-                print "Applying Event Sharing on Group %s" % (g['name'])
-                data = json.dumps({'group_id': int(g['id']), 'what': "TICEVT", 'who': "BIT9"})
-                requests.post(urlparse.urljoin(cb_server_url, '/api/v1/group/{0}/datasharing'.format(g['id'])),
+                if test:
+                    print "%s Group not configured for Sharing Events" % (g['name'])
+                else: 
+                    print "Applying Event Sharing on Group %s" % (g['name'])
+                    data = json.dumps({'group_id': int(g['id']), 'what': "TICEVT", 'who': "BIT9"})
+                    requests.post(urlparse.urljoin(cb_server_url, '/api/v1/group/{0}/datasharing'.format(g['id'])),
                                       data=data, headers={'X-Auth-Token': api_token}, verify=ssl_verify)
 
-def change_config_setting(setting, value, config):
+def change_config_setting(setting, value, config,test=None):
+
+    if test:
+        print "----Validating configuration against %s----" % (config)
     found = False
     update = False
     fh, abs_path = tempfile.mkstemp()
@@ -142,9 +183,12 @@ def change_config_setting(setting, value, config):
                 if line.lstrip().startswith('{0}='.format(setting)) or line.lstrip().startswith('#{0}='.format(setting)):
                     found = True
                     if str(value).strip() != str(line.strip().split("=")[1]) or line.lstrip().startswith('#{0}='.format(setting)):
-                        print 'Changing %s setting [setting=%s][value=%s]' % (config, setting, str(value))
-                        new_file.write('{0}={1}\n'.format(setting, value.strip()))
-                        update=True
+                        if test:
+                            print 'Incorrect Cb.conf option %s setting [setting=%s][value=%s]' % (config, setting, str(value))
+                        else:
+                            print 'Changing %s setting [setting=%s][value=%s]' % (config, setting, str(value))
+                            new_file.write('{0}={1}\n'.format(setting, value.strip()))
+                            update=True
                     else:
                        new_file.write(line) 
                 else:
@@ -157,15 +201,18 @@ def change_config_setting(setting, value, config):
     if update==False:
         os.remove(abs_path)
     else:
-        os.remove(config)
-        shutil.move(abs_path, config)
-        subprocess.call(['/bin/chown','root:cb', config])
-        subprocess.call(['/bin/chmod','644' ,config])
+        if test:
+            os.remove(abs_path)
+        else:
+            os.remove(config)
+            shutil.move(abs_path, config)
+            subprocess.call(['/bin/chown','root:cb', config])
+            subprocess.call(['/bin/chmod','644' ,config])
 
 
-def update_cb_conf(config):
+def update_cb_conf(config, test=None):
     for k,v in config.items():
-        change_config_setting(k, v, '/etc/cb/cb.conf')
+        change_config_setting(k, v, '/etc/cb/cb.conf',test)
 
 def get_apitoken(cb_server_url, username, password, ssl_verify=True):
     if cb_server_url.endswith('/'):
@@ -191,13 +238,15 @@ def build_cli_parser():
                       help="API Token for Carbon Black server")
     parser.add_option("-n", "--no-ssl-verify", action="store_false", default=True, dest="ssl_verify",
                       help="Do not verify server SSL certificate.")
-    parser.add_option("-f", "--file", action="store", default=True, dest="buildfile",
+    parser.add_option("-f", "--file", action="store", default=False, dest="buildfile",
                       help="Configuration.ini file that contains the configuration to be applied")   
     parser.add_option("-r", "--restart", action="store_true", default=False, dest="servicerestart",
                       help="Restart Cb-Enterpise Services upon completion of script.  If applying any cb.conf changes a Carbon Black Service restart will be required")                        
     parser.add_option("--cbinit", action="store_true", default=False, dest="cbinit",
                       help="Execute the command /usr/share/cb/cbinit with the options from the configuration file")    
-
+    parser.add_option("-t", "--test", action="store_true", default=False, dest="test",
+                      help="Do NOT apply any configuration changes only test and report")
+    
     return parser
     
 def main(argv):
@@ -206,7 +255,7 @@ def main(argv):
     if not opts.buildfile:
         print "Missing required param; run with --help for usage"
         sys.exit(-1)
-    if opts.cbinit:
+    if opts.cbinit and not opts.test:
         if os.path.isfile('/etc/cb/server.token'):
             os.remove('/etc/cb/server.token')
         print "Re-Initializing Carbon Black"
@@ -219,15 +268,17 @@ def main(argv):
         api_token = opts.token
         api_server = opts.server_url
     feedconfig = load(opts.buildfile,section='Feed')
-    enable_feed_action(api_server, api_token, feedconfig , opts.ssl_verify)
+    enable_feed_action(api_server, api_token, feedconfig , opts.ssl_verify, opts.test)
     sharingconfig = load(opts.buildfile,section='Sharing')
-    enable_data_sharing(api_server, api_token, sharingconfig, opts.ssl_verify )
+    enable_data_sharing(api_server, api_token, sharingconfig, opts.ssl_verify, opts.test )
     groupconfig = load(opts.buildfile,section='Group')
-    configure_groups(api_server, api_token, groupconfig, opts.ssl_verify )
+    configure_groups(api_server, api_token, groupconfig, opts.ssl_verify, opts.test )
+    if opts.test:
+        report_watchlist_action(api_server, api_token, opts.ssl_verify)
     if os.path.isfile('/etc/cb/cb.conf'):
         cb_conf_config = load(opts.buildfile,section='cb.conf')
-        update_cb_conf(cb_conf_config )
-        if opts.servicerestart:
+        update_cb_conf(cb_conf_config, opts.test )
+        if opts.servicerestart and not opts.test:
             print "Restarting Carbon Black Services"
             subprocess.call(['/etc/init.d/cb-enterprise', 'restart'])
